# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MAWIBO Super Health Ecosystem API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List doctors with filters
      parameters:
        - name: specialty
          in: query
          schema:
            type: string
        - name: county
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get doctor details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
...
  /hospitals:
    get:
      operationId: listHospitals
      tags: [hospitals]
      summary: List hospitals and clinics
      parameters:
        - name: county
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Hospitals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hospital"
  /hospitals/{id}:
    get:
      operationId: getHospital
      tags: [hospitals]
      summary: Get hospital details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hospital details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"

  # ─── PHARMACIES ───────────────────────────────────────────────────────────
...
  /labs/bookings:
    get:
      operationId: listLabBookings
      tags: [labs]
      summary: List lab test bookings
      responses:
        "200":
          description: Lab bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabBooking"
    post:
      operationId: createLabBooking
      tags: [labs]
      summary: Book a lab test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabBookingInput"
      responses:
        "201":
          description: Lab booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabBooking"

  # ─── MENTAL HEALTH ────────────────────────────────────────────────────────
...
  /emergency/contacts:
    get:
      operationId: listEmergencyContacts
      tags: [emergency]
      summary: List emergency contacts
      responses:
        "200":
          description: Emergency contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyContact"
    post:
      operationId: createEmergencyContact
      tags: [emergency]
      summary: Add emergency contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyContactInput"
      responses:
        "201":
          description: Emergency contact added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContact"
  /emergency/contacts/{id}:
    delete:
      operationId: deleteEmergencyContact
      tags: [emergency]
      summary: Remove emergency contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── BLOOD BANK ───────────────────────────────────────────────────────────
...
  /dashboard/vitals:
    get:
      operationId: getVitalsTrend
      tags: [dashboard]
      summary: Get vitals trend data
      responses:
        "200":
          description: Vitals data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalsTrend"
    post:
      operationId: logVitals
      tags: [dashboard]
      summary: Log today's vitals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VitalsInput"
      responses:
        "201":
          description: Vitals logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalsEntry"
  /dashboard/water-tracker:
    get:
      operationId: getWaterTracker
      tags: [dashboard]
      summary: Get today's water intake
      responses:
        "200":
          description: Water intake
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterTracker"
    post:
      operationId: logWaterIntake
      tags: [dashboard]
      summary: Log water intake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaterLogInput"
      responses:
        "200":
          description: Updated water tracker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterTracker"
components:
  schemas:
...
    Doctor:
      type: object
      required: [id, name, specialty, hospital, county, rating, consultationFee, available, imageUrl]
      properties:
        id: { type: integer }
        name: { type: string }
        specialty: { type: string }
        hospital: { type: string }
        county: { type: string }
        city: { type: string }
        rating: { type: number }
        reviewCount: { type: integer }
        consultationFee: { type: number }
        available: { type: boolean }
        imageUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        yearsExperience: { type: integer }
        languages: { type: array, items: { type: string } }
...
    Hospital:
      type: object
      required: [id, name, type, county, city, phone, rating]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        county: { type: string }
        city: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        address: { type: string }
        rating: { type: number }
        reviewCount: { type: integer }
        services: { type: array, items: { type: string } }
        departments: { type: array, items: { type: string } }
        imageUrl: { type: ["string", "null"] }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        emergencyAvailable: { type: boolean }
        openNow: { type: boolean }
...
    LabBooking:
      type: object
      required: [id, labId, labName, testName, date, status]
      properties:
        id: { type: integer }
        labId: { type: integer }
        labName: { type: string }
        testName: { type: string }
        date: { type: string }
        time: { type: string }
        status: { type: string }
        homeCollection: { type: boolean }
        resultUrl: { type: ["string", "null"] }
    LabBookingInput:
      type: object
      required: [labId, testName, date, time]
      properties:
        labId: { type: integer }
        testName: { type: string }
        date: { type: string }
        time: { type: string }
        homeCollection: { type: boolean }
...
    EmergencyContact:
      type: object
      required: [id, name, phone, relationship]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        relationship: { type: string }
    EmergencyContactInput:
      type: object
      required: [name, phone, relationship]
      properties:
        name: { type: string }
        phone: { type: string }
        relationship: { type: string }
...
    VitalsTrend:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/VitalsEntry"
    VitalsEntry:
      type: object
      required: [id, date]
      properties:
        id: { type: integer }
        date: { type: string }
        systolic: { type: ["integer", "null"] }
        diastolic: { type: ["integer", "null"] }
        heartRate: { type: ["integer", "null"] }
        bloodSugar: { type: ["number", "null"] }
        temperature: { type: ["number", "null"] }
        oxygenSaturation: { type: ["integer", "null"] }
        weight: { type: ["number", "null"] }
    VitalsInput:
      type: object
      properties:
        systolic: { type: integer }
        diastolic: { type: integer }
        heartRate: { type: integer }
        bloodSugar: { type: number }
        temperature: { type: number }
        oxygenSaturation: { type: integer }
        weight: { type: number }
    WaterTracker:
      type: object
      required: [date, consumed, goal, percentage]
      properties:
        date: { type: string }
        consumed: { type: number }
        goal: { type: number }
        percentage: { type: number }
        logs:
          type: array
          items:
            type: object
            properties:
              amount: { type: number }
              time: { type: string }
    WaterLogInput:
      type: object
      required: [amount]
      properties:
        amount: { type: number }